+++ src/pages/PasswordPage.tsx
@@ -7,7 +7,6 @@ import {
 import { useNavigate } from "react-router-dom";
 
 import { css } from "@emotion/react";
-import { useQueryClient } from "@tanstack/react-query";
 
 import Button from "@components/atoms/Button";
 import Flex from "@components/atoms/Flex";
@@ -36,7 +35,6 @@ type passwordFormState = {
 const PasswordPage = () => {
   const userPasswordUpdate = useUserPasswordUpdateMutation();
   const theme = useThemeStore((state) => state.theme);
-  const queryClient = useQueryClient();
   const navigate = useNavigate();
   const [form, setForm] = useState<passwordFormState>({
     password: "",
@@ -54,7 +52,7 @@ const PasswordPage = () => {
 
   const handleUpdatePassword = () => {
     if (form.password === "" || form.passwordConfirm === "") {
-      scrawlToast.error("비밀번호 혹은 비밀번호 확인이 입력되지 않았습니다.");
+      scrawlToast.error("모든 입력이 완료되지 않았습니다.");
       return;
     }
     if (form.password !== form.passwordConfirm) {
@@ -65,16 +63,7 @@ const PasswordPage = () => {
       scrawlToast.error("패스워드 형식이 맞지 않습니다.");
       return;
     }
-
-    try {
-      userPasswordUpdate.mutate(form);
-      queryClient.clear();
-      navigate(DOMAIN.HOME);
-      scrawlToast.success("비빌번호 변경에 성공했습니다.");
-    } catch (e) {
-      scrawlToast.error("비밀번호 변경 중 오류가 발생하였습니다.");
-      return;
-    }
+    userPasswordUpdate.mutate(form);
   };
 
   const handleUpdatePasswordWithEnter: KeyboardEventHandler<HTMLElement> = (
